[archimedes] Move `load_json` to utils file
This code converts the `load_json` method in the manager class to a
function and moves it to the utils file.
